+++ android/src/main/java/com/reactnativestripesdk/utils/Mappers.kt
@@ -314,7 +314,7 @@ internal fun mapFromCard(card: Card?): WritableMap? {
 internal fun mapFromToken(token: Token): WritableMap {
   val tokenMap: WritableMap = WritableNativeMap()
   tokenMap.putString("id", token.id)
-  tokenMap.putInt("created", token.created.time.toInt())
+  tokenMap.putDouble("created", token.created.time.toDouble())
   tokenMap.putString("type", mapTokenType(token.type))
   tokenMap.putBoolean("livemode", token.livemode)
   tokenMap.putMap("bankAccount", mapFromBankAccount(token.bankAccount))
